+++ app/src/renderer/appRenderer.tsx
@@ -7,13 +7,10 @@ import { inDev } from '@common/helpers';
 import { CssBaseline, ThemeProvider } from '@mui/material';
 
 import { theme } from '@renderer/config';
-import { CacheSystem } from '@renderer/api/CacheSystem';
 import App from './App';
 
 import { Provider } from 'react-redux';
-import { store } from '@renderer/state/store';
-
-CacheSystem.Clear();
+import { store } from '@renderer/store';
 
 // Render application in DOM
 ReactDOM.render(
