+++ app/src/renderer/hooks/index.ts
@@ -1,5 +1,3 @@
 export { useCache } from './useCache';
 export { useSnackbar } from './useSnackbar';
 export { useSwitch } from './useSwitch';
-
-export { useAppDispatch, useAppSelector } from './useRedux';
