--- app/src/renderer/hooks/useRedux.ts
@@ -1,7 +1,0 @@
-// eslint-disable-next-line import/named
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from '@renderer/state/store';
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
