--- app/src/renderer/state/reducers/index.ts
@@ -1,1 +1,0 @@
-export { dataManagerReducer } from './dataManagerReducer';
