+++ app/src/renderer/store/data-manager/types.ts
